fixed #66
