+++ lib/app/pages/splash/splash_page.dart
@@ -1,7 +1,5 @@
 import 'package:corona_data/app/pages/splash/splash_controller.dart';
-import 'package:corona_data/app/shared/utils/constants.dart';
-import 'package:corona_data/app/shared/widgets/animations/splash_virus_animation.dart';
-import 'package:corona_data/app/shared/widgets/animations/virus_circular_animation.dart';
+import 'package:corona_data/app/shared/widgets/animations/virus_infection_animation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
@@ -16,15 +14,14 @@ class SplashPage extends StatefulWidget {
 class _SplashPageState extends ModularState<SplashPage, SplashController> {
   @override
   Widget build(BuildContext context) {
-    final size = MediaQuery.of(context).size;
+    // final size = MediaQuery.of(context).size;
     return Scaffold(
       backgroundColor: Theme.of(context).primaryColorDark,
       body: Center(
-          child: SplashVirusAnimation(
-        animation: SplashVirusAnimations.dangerous,
+          child: VirusInfectionCircularAnimation(
+        animation: VirusInfectionAnimation.default_animation,
         fit: BoxFit.contain,
-        width: size.width,
-        heigth: size.height,
+        size: 400,
       )),
     );
   }
